Sam/server_side/requirements.txt

python-crontab
psycopg2
bs4
requests
